store parsed ip from NetworkAddress as a tuple
We don't need to use that binary format anywhere outside this message serialization, so the format used by erlang seems handy - it's simple to store and it's accepted by things like :gen_tcp. An alternative would be to keep a string representation but then it takes more space in memory and we would need to convert it to something more compact e.g. when storing list of known addrs on disk.